pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contextcraft"
version = "0.1.0"
description = "Analyzes a code repository and generates a structured Context Pack for AI coding tools"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "ContextCraft" }]
keywords = ["cli", "code-analysis", "ai", "context-pack"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "anthropic>=0.18.0",
    "gitpython>=3.1.0",
    "tree-sitter>=0.20.0",
    "tree-sitter-python>=0.20.0",
    "tree-sitter-javascript>=0.20.0",
    "tree-sitter-typescript>=0.20.0",
    "tree-sitter-java>=0.20.0",
    "pathspec>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
contextcraft = "contextcraft.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["contextcraft*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
